+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## [Unreleased]
 
+## NumericTextBox
+### Bug Fixes
+* Removed height property from the NumericTextBox.
+
 ## 1.0.11-preview (2017-08-03)
 
 ## TextBox